using about components from bento-tools (#231)
About page from bento components
