+++ src/pages/about/aboutController.js
@@ -2,8 +2,8 @@ import React, { useState, useEffect } from 'react';
 import yaml from 'js-yaml';
 import axios from 'axios';
 import YAMLData from '../../content/prod/aboutPagesContent.yaml';
-import AboutBody from './aboutBodyView';
 import env from '../../utils/env';
+import AboutView from './aboutView';
 
 const ABOUT_CONTENT_URL = env.REACT_APP_ABOUT_CONTENT_URL;
 
@@ -30,17 +30,7 @@ const About = ({ match }) => {
   }, [match.path]);
 
   return (
-    <>
-      <AboutBody data={{
-        img: data.primaryContentImage ? data.primaryContentImage : '',
-        title: data.title ? data.title : '',
-        content: data.content ? data.content : '',
-        table: data.table ? data.table : '',
-        secondaryZoomImage: data.secondaryZoomImage ? data.secondaryZoomImage : null,
-        secondaryZoomImageTitle: data.secondaryZoomImageTitle ? data.secondaryZoomImageTitle : null,
-      }}
-      />
-    </>
+    <AboutView data={data} />
   );
 };
 export default About;
