+++ src/pages/modelPage/modelPageView.js
@@ -8,7 +8,7 @@ import yaml from 'js-yaml';
 import axios from 'axios';
 import YAMLData from '../../content/prod/aboutPagesContent.yaml';
 import stockImg from '../../assets/about/About_CTDC_Model.png';
-import AboutBody from '../about/aboutBodyView';
+import AboutBody from '../about/aboutView';
 import CenterIcon from '../../assets/about/Model-Buttons-Center.svg';
 import ZoomInIcon from '../../assets/about/Model-Buttons-ZoomIn.svg';
 import ZoomOutIcon from '../../assets/about/Model-Buttons-ZoomOut.svg';
